+++ apps/sample-imdb/app/search/[term]/page.tsx
@@ -1,0 +1,12 @@
+import { SearchResult } from '~/app/search-result'
+
+export default async function Movie({ params }: { params: { term: string } }) {
+  const { term } = params
+
+  return (
+    <>
+      {/** @ts-expect-error */}
+      <SearchResult searchTerm={term} />
+    </>
+  )
+}
