+++ apps/sample-imdb/components/search.tsx
@@ -15,7 +15,7 @@ export const Search = ({ term }: { term: string }) => {
         ) as string
 
         if (value.length) {
-          router.push(`/?searchTerm=${value}`)
+          router.push(`/search/${value}`)
         }
       }}
     >
@@ -25,7 +25,7 @@ export const Search = ({ term }: { term: string }) => {
         defaultValue={term}
         onChange={debounce((evt: ChangeEvent<HTMLInputElement>) => {
           if (evt.target.value) {
-            router.push(`/?searchTerm=${evt.target.value}`)
+            router.push(`/search/${evt.target.value}`)
           }
         }, 300)}
         className="h-[2em] w-[50ch] pl-4 border-2 border-neutral-600 rounded-2xl bg-transparent text-white outline-none focus:transform focus:scale-110 origin-left transition-all focus:border-pink-500"
